requirements.txt
Django==5.2
psycopg2-binary==2.9.10
social-auth-app-django==5.4.3
djangorestframework==3.16.0
django-cors-headers==4.7.0
PyJWT==2.10.1
cryptography==44.0.2
defusedxml==0.7.1
oauthlib==3.2.2
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
python-decouple==3.9
django-allauth==0.61.0
PyPDF2==3.0.1
openai==1.17.0 